requirements.txt
django-cors-headers==4.6.0
django-environ==0.12.0
django-modeltranslation==0.19.12
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.8
pdfkit==1.0.0
pillow==11.1.0
pip-chill==1.0.3
pymysql==1.1.1
sendgrid-django==4.2.0
stripe==11.6.0
